+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - Forward logs to Kafka directly instead of serialized as envelope. ([#4875](https://github.com/getsentry/relay/pull/4875))
 - Add `gen_ai.response.tokens_per_second` span attribute on AI spans. ([#4883](https://github.com/getsentry/relay/pull/4883))
+- Add support for playstation data requests. ([#4870](https://github.com/getsentry/relay/pull/4870))
 
 ## 25.6.2