+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Preserve user specified event values in Unreal crash reports. ([#4882](https://github.com/getsentry/relay/pull/4882))
 - OS name parsing of Unreal crash reports. ([#4854](https://github.com/getsentry/relay/pull/4854))
+- Do not overwrite geo information if already set. ([#4888](https://github.com/getsentry/relay/pull/4888))
 
 **Internal**: